--- src/SrpTask/Game/IEnemy.cs
@@ -1,7 +1,0 @@
-﻿namespace SrpTask.Game
-{
-    public interface IEnemy
-    {
-        void TakeDamage(int amount);
-    }
-}
