+++ src/SrpTask/IEnemy.cs
@@ -1,0 +1,7 @@
+﻿namespace SrpTask
+{
+    public interface IEnemy
+    {
+        void TakeDamage(int amount);
+    }
+}
