+++ src/SrpTask/IGameEngine.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace SrpTask
+{
+    public interface IGameEngine
+    {
+        void PlaySpecialEffect(string effectName);
+        List<IEnemy> GetEnemiesNear(RpgPlayer player);
+    }
+}
